Updated styles watch task to watch any symlinked CF components
When updating a component in the capital-framework repo this project could not detect the change and users had to kick off the build process manually. Updated the watch task to include each less file from each component so that changes can be detected. It still requires a manual browser refresh, but that seems to be an outside issue.